sped up fetching number of files with concurrency
using a ThreadPoolExecutor, we can speed up the way cnda-dl fetches the number of files associated
with a session by using concurrent database connections (which, after trial-and-error, turns out
to be 10 connections at a time. this is added as a constant at the top of the script). this improved
the speed of one test session from 32 seconds to 6 seconds!